+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.5.1, 25 July 2026
 
 * Fixed bug in dtms_censoring which led to gaps being not flagged correctly. Thanks to Andreas Ljungström (https://github.com/andreasljungstrm) for finding and fixing. 
+* Added variance matrix to dtms_nonparametric using binomial approximation
 
 # Version 0.5.0, 22 July 2026